Änderungen für Datensatz mit 100000 Strecken, Datentyp von Float zu Double geändert

+++ Abgabe 3/PlaneSweep/src/IntersectingEdgeChecker.java
@@ -21,14 +21,14 @@ public class IntersectingEdgeChecker {
 		    while(zeile != null) {
 		    	String[] zeilenArray = zeile.split(" ");
 		    	//Die die links anfangen sind die Startpunkte
-		    	float x1 = Float.parseFloat(zeilenArray[0]);
-		    	float x2 = Float.parseFloat(zeilenArray[2]);
+		    	double x1 = Double.parseDouble(zeilenArray[0]);
+		    	double x2 = Double.parseDouble(zeilenArray[2]);
 		    	Strecke s;
 		    	if(x1 <= x2) {
-		    		s = new Strecke(x1,Float.parseFloat(zeilenArray[1]),x2,Float.parseFloat(zeilenArray[3]));
+		    		s = new Strecke(x1,Double.parseDouble(zeilenArray[1]),x2,Double.parseDouble(zeilenArray[3]));
 		    	}
 		    	else {
-		    		s = new Strecke(x2,Float.parseFloat(zeilenArray[3]),x1,Float.parseFloat(zeilenArray[1]));
+		    		s = new Strecke(x2,Double.parseDouble(zeilenArray[3]),x1,Double.parseDouble(zeilenArray[1]));
 		    	}	    	
 		    	Strecken.add(s);
 		    	zeile = br.readLine();
@@ -44,8 +44,8 @@ public class IntersectingEdgeChecker {
 		return Strecken;
 	}
 	
-	public float ccw(Point2D.Float p1, Point2D.Float p2, Point2D.Float p3) {
-		float edge[] = new float[3];
+	public double ccw(Point2D.Double p1, Point2D.Double p2, Point2D.Double p3) {
+		double edge[] = new double[3];
 		edge[0] = (p2.x - p1.x) * (p2.y + p1.y);
 		edge[1] = (p3.x - p2.x) * (p3.y + p2.y);
 		edge[2] = (p1.x - p3.x) * (p1.y + p3.y);
@@ -54,8 +54,8 @@ public class IntersectingEdgeChecker {
 	
 	public int doIntersect(Strecke s1, Strecke s2) {
 		//kolinear und überlapppend
-			float test = ccw(s1.getStartPoint(), s1.getEndPoint(),s2.getStartPoint());
-			float test1 = ccw(s1.getStartPoint(), s1.getEndPoint(),s2.getEndPoint());
+			double test = ccw(s1.getStartPoint(), s1.getEndPoint(),s2.getStartPoint());
+			double test1 = ccw(s1.getStartPoint(), s1.getEndPoint(),s2.getEndPoint());
 		if (ccw(s1.getStartPoint(), s1.getEndPoint(),s2.getStartPoint()) == 0 && ccw(s1.getStartPoint(), s1.getEndPoint(),s2.getEndPoint()) == 0) {
 			return 0;	
 		}
@@ -68,7 +68,7 @@ public class IntersectingEdgeChecker {
 		return -1;
 	}
 	
-	public Point2D.Float getIntersectionPoint(Strecke s1, Strecke s2) {
+	public Point2D.Double getIntersectionPoint(Strecke s1, Strecke s2) {
 		//Tausch falls s1 Linie mit nur einem Punkt ist
 		/*if(s1.getStartPoint().equals(s1.getEndPoint())) {
 			Strecke swap = s1;
@@ -77,13 +77,13 @@ public class IntersectingEdgeChecker {
 		}*/
 		int i = doIntersect(s1,s2);
 		if(doIntersect(s1,s2) == 1) {
-			Point2D.Float p1 = s1.getStartPoint();
-			Point2D.Float p2 = s1.getEndPoint();
-			Point2D.Float p3 = s2.getStartPoint();
-			Point2D.Float p4 = s2.getEndPoint();
-			float xs = (float) (((p4.getX() - p3.getX()) * (p2.getX()*p1.getY() - p1.getX()*p2.getY()) - (p2.getX() - p1.getX()) * (p4.getX()*p3.getY() - p3.getX()*p4.getY())) / ((p4.getY() - p3.getY()) * (p2.getX() - p1.getX()) - (p2.getY() - p1.getY()) * (p4.getX() - p3.getX())));
-			float ys = (float) (((p1.getY() - p2.getY()) * (p4.getX()*p3.getY() - p3.getX()*p4.getY()) - (p3.getY() - p4.getY()) * (p2.getX()*p1.getY() - p1.getX()*p2.getY())) / ((p4.getY() - p3.getY()) * (p2.getX() - p1.getX()) - (p2.getY() - p1.getY()) * (p4.getX() - p3.getX())));
-			return new Point2D.Float(xs, ys);
+			Point2D.Double p1 = s1.getStartPoint();
+			Point2D.Double p2 = s1.getEndPoint();
+			Point2D.Double p3 = s2.getStartPoint();
+			Point2D.Double p4 = s2.getEndPoint();
+			double xs = (double) (((p4.getX() - p3.getX()) * (p2.getX()*p1.getY() - p1.getX()*p2.getY()) - (p2.getX() - p1.getX()) * (p4.getX()*p3.getY() - p3.getX()*p4.getY())) / ((p4.getY() - p3.getY()) * (p2.getX() - p1.getX()) - (p2.getY() - p1.getY()) * (p4.getX() - p3.getX())));
+			double ys = (double) (((p1.getY() - p2.getY()) * (p4.getX()*p3.getY() - p3.getX()*p4.getY()) - (p3.getY() - p4.getY()) * (p2.getX()*p1.getY() - p1.getX()*p2.getY())) / ((p4.getY() - p3.getY()) * (p2.getX() - p1.getX()) - (p2.getY() - p1.getY()) * (p4.getX() - p3.getX())));
+			return new Point2D.Double(xs, ys);
 		}
 		//Falls kolinear überlappend(z.B. bei Linie die nur Punkt ist)
 		if(doIntersect(s1,s2) == 0) {
